+++ packages/sdks/src/blocks/form/form/form.lite.tsx
@@ -77,6 +77,9 @@ export default function FormComponent(props: FormProps) {
     submissionState(): FormState {
       return (isEditing() && props.previewState) || state.formState;
     },
+    errorResponse(response: any) {
+      return JSON.stringify(response, null, 2);
+    },
     onSubmit(event: any) {
       const sendWithJsProp =
         props.sendWithJs || props.sendSubmissionsTo === 'email';
@@ -378,7 +381,7 @@ export default function FormComponent(props: FormProps) {
           class="builder-form-error-text"
           css={{ padding: '10px', color: 'red', textAlign: 'center' }}
         >
-          {JSON.stringify(state.responseData, null, 2)}
+          {state.errorResponse(state.responseData)}
         </pre>
       </Show>