+++ packages/sdks/output/angular/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @builder.io/sdk-angular
 
+## 0.21.0
+
+### Minor Changes
+
+- 774b180: 🚨 Breaking Change
+  Refactor Angular SDK to leverage Angular v17+ features (signals, computed, inputs, declarative statements, etc.).
+
+  - Performance revamp: Recomputations now occur only when a dependent signal updates, instead of on every change detection cycle.
+  - Components with children no longer re-render on each change, eliminating sluggish interactions in the Visual Editor.
+
+  The SDK will only work for apps with Angular v17.3.0 and above
+
 ## 0.20.1
 
 ### Patch Changes